+++ pkgs/default.nix
@@ -14,5 +14,6 @@
   bravais-mcp = pkgs.callPackage ./bravais-mcp/package.nix { };
   opencode-desktop = pkgs.callPackage ./opencode-desktop/package.nix { };
   goose-desktop = pkgs.callPackage ./goose-desktop/package.nix { };
+  adguardvpn-cli = pkgs.callPackage ./adguardvpn-cli/package.nix { };
 }